Soil/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
reportlab==4.0.7
qrcode==7.4.2
python-dotenv==1.0.0